# src/docs/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SellMate API",
    "version": "1.0.0",
    "description": "\n# SellMate Platform - Social Commerce Automation Backend API\n\nSellMate is an intelligent social commerce automation platform designed specifically for Nigerian micro-entrepreneurs who sell through Instagram and WhatsApp.\n\n## Features\n- **Unified Conversations**: Consolidate WhatsApp and Instagram messages in one place\n- **AI-Powered Order Detection**: Automatically detect orders from conversations\n- **Profit Analytics**: Real-time business metrics and profit calculations\n- **Customer Management**: Track customer interactions across platforms\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n## Rate Limiting\nAPI requests are rate-limited to prevent abuse. Current limits:\n- General: 100 requests per 15 minutes\n- Authentication: 5 requests per 15 minutes\n- AI endpoints: 20 requests per minute\n      ",
    "contact": {
      "name": "SellMate Team",
      "email": "support@sellmate.ng"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.sellmate.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /auth/verify-otp or /auth/refresh"
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                },
                "message": {
                  "type": "string",
                  "example": "Invalid or expired token"
                }
              }
            }
          }
        }
      },
...
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/oauth/whatsapp": {
      "get": {
        "summary": "Initiate WhatsApp OAuth flow",
        "description": "Initiates the OAuth 2.0 flow for connecting a WhatsApp Business account.\nReturns an authorization URL that the user should be redirected to.\nThe user will authenticate with Meta and grant permissions to SellMate.\n",
        "tags": [
          "OAuth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth initiation successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "authorizationUrl": {
                          "type": "string",
                          "format": "uri",
                          "description": "URL to redirect the user for OAuth authorization"
                        },
                        "state": {
                          "type": "string",
                          "description": "State parameter for CSRF protection"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/v1/auth/oauth/whatsapp/callback": {
      "get": {
        "summary": "Handle WhatsApp OAuth callback",
        "description": "Handles the OAuth callback from Meta after user authorization.\nExchanges the authorization code for access tokens and stores them securely.\nThis endpoint is called by Meta's OAuth server, not directly by the client.\n",
        "tags": [
          "OAuth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Meta"
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "State parameter for CSRF validation"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to frontend with success or error status"
          },
          "400": {
            "description": "Invalid state or authorization code"
          }
        }
      }
    },
    "/api/v1/auth/oauth/whatsapp/disconnect": {
      "delete": {
        "summary": "Disconnect WhatsApp account",
        "description": "Disconnects the WhatsApp Business account from the user's SellMate account.\nRevokes stored tokens and removes the integration. Existing conversations are preserved.\n",
        "tags": [
          "OAuth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "WhatsApp disconnected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "WhatsApp account disconnected successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "No WhatsApp integration found"
          }
        }
      }
    },
    "/api/v1/auth/oauth/instagram": {
      "get": {
        "summary": "Initiate Instagram OAuth flow",
        "description": "Initiates the OAuth 2.0 flow for connecting an Instagram Business account.\nReturns an authorization URL that the user should be redirected to.\nRequires an Instagram Business or Creator account linked to a Facebook Page.\n",
        "tags": [
          "OAuth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth initiation successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "authorizationUrl": {
                          "type": "string",
                          "format": "uri",
                          "description": "URL to redirect the user for OAuth authorization"
                        },
                        "state": {
                          "type": "string",
                          "description": "State parameter for CSRF protection"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/v1/auth/oauth/instagram/callback": {
      "get": {
        "summary": "Handle Instagram OAuth callback",
        "description": "Handles the OAuth callback from Meta after user authorization.\nExchanges the authorization code for access tokens and stores them securely.\nThis endpoint is called by Meta's OAuth server, not directly by the client.\n",
        "tags": [
          "OAuth"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Meta"
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "State parameter for CSRF validation"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to frontend with success or error status"
          },
          "400": {
            "description": "Invalid state or authorization code"
          }
        }
      }
    },
    "/api/v1/auth/oauth/instagram/disconnect": {
      "delete": {
        "summary": "Disconnect Instagram account",
        "description": "Disconnects the Instagram Business account from the user's SellMate account.\nRevokes stored tokens and removes the integration. Existing conversations are preserved.\n",
        "tags": [
          "OAuth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Instagram disconnected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Instagram account disconnected successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "description": "No Instagram integration found"
          }
        }
      }
    },
...
  }
}